Revert "Add pressure plate"
