--- .vsconfig
@@ -1,6 +1,0 @@
-{
-  "version": "1.0",
-  "components": [
-    "Microsoft.VisualStudio.Workload.ManagedGame"
-  ]
-}
