+++ Assets/Scenes/SampleScene.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 682a85066492a424281b0b28d9c55db2
-PrefabImporter:
+guid: 9fc0d4010bbf28b4594072e72b8655ab
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
