--- Assets/Scenes/Scene1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5231e14e23d6ff047a68b8ce1678981d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
