+++ Assets/Scripts/Human.cs
@@ -1,11 +1,11 @@
 using UnityEngine;
 
-public class Robot : MonoBehaviour
+public class Human : MonoBehaviour
 {
     [SerializeField] private float moveSpeed = 15f;
     [SerializeField] private float jumpForce = 10f;
     private Rigidbody2D rb;
-    [SerializeField] private Transform groundCheck;
+    // Need to be change for isGounded if  double jump
     private bool isGrounded;
     private const string GROUND = "Ground";
 
@@ -20,27 +20,34 @@ public class Robot : MonoBehaviour
     }
 
     private void HandleMovement() {
-        
+
         float moveInput = 0f;
 
-        if (Input.GetKey(KeyCode.LeftArrow)) {
+        if (Input.GetKey(KeyCode.A)) {
             moveInput = -1f; // Move left
         } 
-        else if (Input.GetKey(KeyCode.RightArrow)) {
+        else if (Input.GetKey(KeyCode.D)) {
             moveInput = 1f; // Move right
         }
 
-        rb.velocity = new Vector2(moveInput * moveSpeed, rb.velocity.y);
+        rb.velocity= new Vector2(moveInput * moveSpeed, rb.velocity.y);
     }
 
     private void HandleJump() {
-    //Casts a ray downward and checks for collider
-    RaycastHit2D hit = Physics2D.Raycast(groundCheck.position, Vector2.down, 0.1f);
 
-    isGrounded = hit.collider != null; //If the ray hits something on ground true
+        if (Input.GetKeyDown(KeyCode.W) && isGrounded)
+        {
+            rb.velocity = new Vector2(rb.velocity.x, jumpForce);
+            isGrounded = false; 
+        }
+    }
 
-    if (Input.GetKeyDown(KeyCode.UpArrow) && isGrounded) {
-        rb.velocity = new Vector2(rb.velocity.x, jumpForce);
+    //Ground Check
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.CompareTag(GROUND))
+        {
+            isGrounded = true;
         }
     }
 }
