--- Assets/Scripts/Objects/PressurePlate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 53e1ed759a6cf524db5dbd8a73d14f5a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
