--- Assets/Scripts/Player/forMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0962ccb2543ec774caa53524f41573e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
