+++ Assets/Scripts/Robot.cs
@@ -1,12 +1,10 @@
 using UnityEngine;
 
-public class Human : MonoBehaviour
+public class Robot : MonoBehaviour
 {
     [SerializeField] private float moveSpeed = 15f;
     [SerializeField] private float jumpForce = 10f;
     private Rigidbody2D rb;
-    // Need to be change for isGounded if  double jump
-    [SerializeField] private Transform groundCheck;
     private bool isGrounded;
     private const string GROUND = "Ground";
 
@@ -21,27 +19,34 @@ public class Human : MonoBehaviour
     }
 
     private void HandleMovement() {
-
+        
         float moveInput = 0f;
 
-        if (Input.GetKey(KeyCode.A)) {
+        if (Input.GetKey(KeyCode.LeftArrow)) {
             moveInput = -1f; // Move left
         } 
-        else if (Input.GetKey(KeyCode.D)) {
+        else if (Input.GetKey(KeyCode.RightArrow)) {
             moveInput = 1f; // Move right
         }
 
-        rb.velocity= new Vector2(moveInput * moveSpeed, rb.velocity.y);
+        rb.velocity = new Vector2(moveInput * moveSpeed, rb.velocity.y);
     }
 
     private void HandleJump() {
-    //Casts a ray downward and checks for collider
-    RaycastHit2D hit = Physics2D.Raycast(groundCheck.position, Vector2.down, 0.1f);
 
-    isGrounded = hit.collider != null; //If the ray hits something on ground true
+        if (Input.GetKeyDown(KeyCode.UpArrow) && isGrounded)
+        {
+            rb.velocity = new Vector2(rb.velocity.x, jumpForce);
+            isGrounded = false; 
+        }
+    }
 
-    if (Input.GetKeyDown(KeyCode.W) && isGrounded) {
-        rb.velocity = new Vector2(rb.velocity.x, jumpForce);
+    //Ground Check
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.CompareTag(GROUND))
+        {
+            isGrounded = true;
         }
     }
 }
