+++ Assets/Scripts/testgit.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class testgit : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        //test test test
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        print "hellow worldo1";
+    }
+}
