+++ ProjectSettings/TagManager.asset
@@ -14,8 +14,8 @@ TagManager:
   - UI
   - HumanCam
   - RobotCam
-  - Player
-  - Objects
+  - 
+  - 
   - 
   - 
   -
